+++ test/brando/blueprints/form_test.exs
@@ -2,57 +2,11 @@ defmodule Brando.Blueprint.FormsTest do
   use ExUnit.Case
   import Brando.Test.Support, only: [strip_spark_metadata: 1]
 
-  defmodule BlocksHiddenOpts do
-    use Brando.Blueprint,
-      application: "Brando",
-      domain: "Tests",
-      schema: "BlocksHiddenOpts",
-      singular: "blocks_hidden_opts",
-      plural: "blocks_hidden_opts",
-      gettext_module: Brando.Gettext
-
-    attributes do
-      attribute :title, :string
-    end
-
-    forms do
-      form do
-        blocks :blocks, hidden: {:title, "hide"}
-
-        tab "Content" do
-          fieldset do
-            input :title, :text
-          end
-        end
-      end
-
-      form :with_fn do
-        blocks :blocks, hidden: &__MODULE__.hide_blocks?/1
-
-        tab "Content" do
-          fieldset do
-            input :title, :text
-          end
-        end
-      end
-    end
-
-    def hide_blocks?(_form), do: false
-  end
-
   test "default form" do
     assert strip_spark_metadata(Brando.BlueprintTest.Project.__form__()) ==
              strip_spark_metadata(%Brando.Blueprint.Forms.Form{
                __identifier__: :default,
                name: :default,
-               blocks: [
-                 %Brando.Blueprint.Forms.Input{
-                   name: :blocks,
-                   type: :blocks,
-                   component: nil,
-                   opts: []
-                 }
-               ],
                tabs: [
                  %Brando.Blueprint.Forms.Tab{
                    fields: [
@@ -140,14 +94,6 @@ defmodule Brando.Blueprint.FormsTest do
                %Brando.Blueprint.Forms.Form{
                  __identifier__: :default,
                  name: :default,
-                 blocks: [
-                   %Brando.Blueprint.Forms.Input{
-                     name: :blocks,
-                     type: :blocks,
-                     component: nil,
-                     opts: []
-                   }
-                 ],
                  tabs: [
                    %Brando.Blueprint.Forms.Tab{
                      fields: [
@@ -234,8 +180,8 @@ defmodule Brando.Blueprint.FormsTest do
   end
 
   test "blocks accepts hidden opts" do
-    default_form = BlocksHiddenOpts.__form__()
-    with_fn_form = BlocksHiddenOpts.__form__(:with_fn)
+    default_form = Brando.TraitTest.Project.__form__()
+    with_fn_form = Brando.TraitTest.Project.__form__(:with_fn)
 
     assert [%Brando.Blueprint.Forms.Input{name: :blocks, opts: opts}] = default_form.blocks
     assert Keyword.get(opts, :hidden) == {:title, "hide"}